+++ CHANGELOG.md
@@ -1,5 +1,7 @@
 # CHANGELOG
 
+* change the `:html` output to return `document.documentElement.innerHTML` instead of previously used `document.body.innerHTML`
+
 ## 1.0.0
 
 * initial version